fix(controllers/tekton): fix delete devops pipeline logic
