+++ controllers/tekton/utils_test.go
@@ -1,0 +1,50 @@
+/*
+Copyright 2020 The KubeSphere Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package tekton
+
+import "testing"
+
+// Unit test for function containsString
+func Test_containsString(t *testing.T) {
+	tests := []struct {
+		name  string
+		slice []string
+		s     string
+		want  bool
+	}{
+		{
+			name:  "String slice does not contain the given string",
+			slice: []string{"a", "bb", "ccc"},
+			s:     "dddd",
+			want:  false,
+		},
+		{
+			name:  "String slice contains the given string",
+			slice: []string{"a", "bb", "ccc"},
+			s:     "bb",
+			want:  true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := containsString(tt.slice, tt.s)
+			if got != tt.want {
+				t.Errorf("containsString() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
